+++ data/KOL_acc_tiktok.txt
@@ -1,9 +1,16 @@
 hoaa.hanassii
 reencyngo
+truongdaitrang1555
 cciinnn
 lethikhanhhuyen190
 _linhhkaa_ 
+lienhuong212
+linhtruong.ld
 cindymiicute
+linhbarbie
+tetelinh
+hoangminhngoc21
+baotramthieu
 vietphuongthoa98
 djmievn
 ngoc.eng
@@ -13,3 +20,16 @@ tai2701
 _b.t.v
 quocthanh0308
 gialong294
+otis.102
+tranvandu
+brianltran
+alihoangduong1409
+phamtuanhai
+mactrungkien
+denisdang
+yoontran.official
+btranofficial
+quoctruong1988
+im.quangdang
+ngohainam96
+ttb88
