+++ log/logging.log
@@ -1,8 +1,19 @@
-2023-06-18 15:37:08,751 Starting new HTTPS connection (1): www.instagram.com:443
-2023-06-18 15:37:09,470 https://www.instagram.com:443 "GET /graphql/query?query_hash=d6f4427fbe92d846298cf93df0b937d3&variables=%7B%7D HTTP/1.1" 200 447
-2023-06-18 15:37:10,501 Starting new HTTPS connection (1): i.instagram.com:443
-2023-06-18 15:37:11,628 https://i.instagram.com:443 "GET /api/v1/users/web_profile_info/?username=hoaminzy_rose HTTP/1.1" 200 36232
-2023-06-18 15:37:16,514 Starting new HTTPS connection (1): www.instagram.com:443
-2023-06-18 15:37:17,272 https://www.instagram.com:443 "GET /graphql/query?query_hash=003056d32c2554def87228bc3fd9668a&variables=%7B%22id%22%3A2078808390%2C%22first%22%3A50%2C%22after%22%3A%22QVFBaDFraTRGd0FIYmN6Ml9oeTBZMFhYOUg4S2o1c1hCYmxrWGxXV1MwaHgwa05nSjVGbUphRjhmMm12dDMwZ2VLN1VDMkgtWTBLdDM2TVdOMlhaVUpTQg%3D%3D%22%7D HTTP/1.1" 401 98
-2023-06-18 15:37:17,728 https://www.instagram.com:443 "GET /graphql/query?query_hash=003056d32c2554def87228bc3fd9668a&variables=%7B%22id%22%3A2078808390%2C%22first%22%3A50%2C%22after%22%3A%22QVFBaDFraTRGd0FIYmN6Ml9oeTBZMFhYOUg4S2o1c1hCYmxrWGxXV1MwaHgwa05nSjVGbUphRjhmMm12dDMwZ2VLN1VDMkgtWTBLdDM2TVdOMlhaVUpTQg%3D%3D%22%7D HTTP/1.1" 401 98
-2023-06-18 15:37:22,325 https://www.instagram.com:443 "GET /graphql/query?query_hash=003056d32c2554def87228bc3fd9668a&variables=%7B%22id%22%3A2078808390%2C%22first%22%3A50%2C%22after%22%3A%22QVFBaDFraTRGd0FIYmN6Ml9oeTBZMFhYOUg4S2o1c1hCYmxrWGxXV1MwaHgwa05nSjVGbUphRjhmMm12dDMwZ2VLN1VDMkgtWTBLdDM2TVdOMlhaVUpTQg%3D%3D%22%7D HTTP/1.1" 401 98
+2023-06-21 20:57:30,926 Starting new HTTPS connection (1): www.instagram.com:443
+2023-06-21 20:57:32,138 https://www.instagram.com:443 "GET /graphql/query?query_hash=d6f4427fbe92d846298cf93df0b937d3&variables=%7B%7D HTTP/1.1" 200 444
+2023-06-21 20:57:34,322 Starting new HTTPS connection (1): i.instagram.com:443
+2023-06-21 20:57:35,626 https://i.instagram.com:443 "GET /api/v1/users/web_profile_info/?username=junpham HTTP/1.1" 200 46048
+2023-06-21 20:57:37,161 Starting new HTTPS connection (1): www.instagram.com:443
+2023-06-21 20:57:39,226 https://www.instagram.com:443 "GET /graphql/query?query_hash=003056d32c2554def87228bc3fd9668a&variables=%7B%22id%22%3A23400892%2C%22first%22%3A50%2C%22after%22%3A%22QVFCM1NjeVp1U0ltc1RRRHBYRUhvU1U3eUpIdGo2ZC1iem5acW5pemlVTlhzRmNOdVRRdTVJa1hlX2xNTkp4S3VHUF84ampnTXFkT055RGtldGg3VTV0Xw%3D%3D%22%7D HTTP/1.1" 200 88546
+2023-06-21 20:57:39,810 Starting new HTTPS connection (1): www.instagram.com:443
+2023-06-21 20:57:41,838 https://www.instagram.com:443 "GET /graphql/query?query_hash=003056d32c2554def87228bc3fd9668a&variables=%7B%22id%22%3A23400892%2C%22first%22%3A50%2C%22after%22%3A%22QVFES2JHVmFfcjhoaXZ6TExqTDhUdkpFVDhaQzd5VUtVVndYdUh0eHZCRFdsdWVFY1lvZnZaampMMFVWd1hHMkJHeVBJQ3N3YkRqdDlDWEVYV0syZmtMOQ%3D%3D%22%7D HTTP/1.1" 200 89443
+2023-06-21 20:57:43,343 Starting new HTTPS connection (1): www.instagram.com:443
+2023-06-21 20:57:46,002 https://www.instagram.com:443 "GET /graphql/query?query_hash=003056d32c2554def87228bc3fd9668a&variables=%7B%22id%22%3A23400892%2C%22first%22%3A50%2C%22after%22%3A%22QVFBOUI4M2VidmNNdjl0TFd2ZGxra25XOHVqa2lVNmhqU2JRVkg2U0xDMTdIQ0RCMlVzcFJqak10X2hsWDJrSERUaC10QlduRF9HZmNwZjFsbXlPTTlPRA%3D%3D%22%7D HTTP/1.1" 200 114591
+2023-06-21 20:57:47,031 Starting new HTTPS connection (1): www.instagram.com:443
+2023-06-21 20:57:48,889 https://www.instagram.com:443 "GET /graphql/query?query_hash=003056d32c2554def87228bc3fd9668a&variables=%7B%22id%22%3A23400892%2C%22first%22%3A50%2C%22after%22%3A%22QVFDUHk2V0EzMmlZcXd6akdjX0NfeXM3YXpjZklMbGZZUVJPRW8ya3Q1TUZPMzhuWWxnT1NBYmpGaUt4NjV5T3hHNUd6czZXLXhqZFE3UnBDblozMFhRNw%3D%3D%22%7D HTTP/1.1" 200 92634
+2023-06-21 20:57:52,147 Starting new HTTPS connection (1): www.instagram.com:443
+2023-06-21 20:57:53,841 https://www.instagram.com:443 "GET /graphql/query?query_hash=003056d32c2554def87228bc3fd9668a&variables=%7B%22id%22%3A23400892%2C%22first%22%3A50%2C%22after%22%3A%22QVFCZ2h5bHY0bWx5T1ZMVnlRRjVrOTg1Mi1Qb3M5X3pxOTJuNXJYYVY3NGxsNFU4ZUFMdXVmNGhXYnVHV01Vb1hiZDVMc2o2VXE4NnY5RU9UOGpwQXRPXw%3D%3D%22%7D HTTP/1.1" 200 75768
+2023-06-21 20:57:56,471 Starting new HTTPS connection (1): www.instagram.com:443
+2023-06-21 20:57:58,044 https://www.instagram.com:443 "GET /graphql/query?query_hash=003056d32c2554def87228bc3fd9668a&variables=%7B%22id%22%3A23400892%2C%22first%22%3A50%2C%22after%22%3A%22QVFBaWlXTDczQVlpVWtBc0RaUzFvQmFlZzFRR0VpYmhZUmpUejBYTktzdWVMLTRMbnRwSTVPek9YTlpyaDJEVFlINmJ0dGtRblVleEtTU3BKZTBiUUdISw%3D%3D%22%7D HTTP/1.1" 200 84092
+2023-06-21 20:58:00,866 Starting new HTTPS connection (1): www.instagram.com:443
+2023-06-21 20:58:02,868 https://www.instagram.com:443 "GET /graphql/query?query_hash=003056d32c2554def87228bc3fd9668a&variables=%7B%22id%22%3A23400892%2C%22first%22%3A50%2C%22after%22%3A%22QVFBMkpFXzJqWlJfS25wT19uVW5mREszbmZKR0tQVkVTZEdPcE1GUDkzWjlDdFZPMldRUlpRalE3V3NuTXlTUW9wLTBTV01HMnFobml4STlRbThKOF81Ug%3D%3D%22%7D HTTP/1.1" 200 105985
+2023-06-21 20:58:03,324 Starting new HTTPS connection (1): www.instagram.com:443
